+++ be/src/vec/exec/format/parquet/bool_rle_decoder.h
@@ -35,6 +35,7 @@ struct Slice;
 } // namespace doris
 
 namespace doris::vectorized {
+#include "common/compile_check_begin.h"
 class BoolRLEDecoder final : public Decoder {
 public:
     BoolRLEDecoder() = default;
@@ -49,7 +50,7 @@ public:
     Status _decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
                           ColumnSelectVector& select_vector, bool is_dict_filter);
 
-    Status skip_values(size_t num_values) override;
+    Status skip_values(int num_values) override;
 
 private:
     RleDecoder<uint8_t> _decoder;
@@ -57,4 +58,6 @@ private:
     size_t _num_bytes;
     size_t _current_value_idx = 0;
 };
+#include "common/compile_check_end.h"
+
 } // namespace doris::vectorized
