+++ be/src/vec/exec/format/parquet/byte_array_plain_decoder.cpp
@@ -25,8 +25,8 @@
 #include "vec/data_types/data_type_nullable.h"
 
 namespace doris::vectorized {
-
-Status ByteArrayPlainDecoder::skip_values(size_t num_values) {
+#include "common/compile_check_begin.h"
+Status ByteArrayPlainDecoder::skip_values(int num_values) {
     for (int i = 0; i < num_values; ++i) {
         if (UNLIKELY(_offset + 4 > _data->size)) {
             return Status::IOError("Can't read byte array length from plain decoder");
@@ -105,4 +105,6 @@ Status ByteArrayPlainDecoder::_decode_values(MutableColumnPtr& doris_column, Dat
     }
     return Status::OK();
 }
+#include "common/compile_check_end.h"
+
 } // namespace doris::vectorized
