+++ be/src/vec/exec/format/parquet/byte_array_plain_decoder.h
@@ -41,7 +41,7 @@ class ColumnDecimal;
 } // namespace doris
 
 namespace doris::vectorized {
-
+#include "common/compile_check_begin.h"
 class ByteArrayPlainDecoder final : public Decoder {
 public:
     ByteArrayPlainDecoder() = default;
@@ -54,6 +54,8 @@ public:
     Status _decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
                           ColumnSelectVector& select_vector, bool is_dict_filter);
 
-    Status skip_values(size_t num_values) override;
+    Status skip_values(int num_values) override;
 };
+#include "common/compile_check_end.h"
+
 } // namespace doris::vectorized
